Moved crafting page patch
